% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_dat}
\alias{get_dat}
\title{Get Bike Data}
\usage{
get_dat(url, which)
}
\arguments{
\item{url}{URL of API}

\item{which}{which data set to get}
}
\value{
data.frame
}
\description{
Get Bike Data
}
\examples{
get_dat("http://127.0.0.1:5249", "station")
}
