% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_feed_dat}
\alias{get_feed_dat}
\title{Get data from a feed}
\usage{
get_feed_dat(feed)
}
\arguments{
\item{feed}{name of a feed}
}
\value{
data from feed
}
\description{
Get data from a feed
}
\examples{
get_feed_dat("station_status")
}
