% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_feed_url}
\alias{get_feed_url}
\title{Get URL of a Feed}
\usage{
get_feed_url(feed)
}
\arguments{
\item{feed}{name of a feed}
}
\value{
list of data and update time
}
\description{
Get URL of a Feed
}
\examples{
get_feed_url("station_information")
}
