% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{make_bike_map}
\alias{make_bike_map}
\title{Make Leaflet map of bike data}
\usage{
make_bike_map(dat, stations)
}
\arguments{
\item{dat}{output from make_plot_dat}
}
\value{
leaflet object
}
\description{
Make Leaflet map of bike data
}
\examples{
dat <- make_plot_dat(pins::pin_get("break_up_station_dat"), pins::pin_get("break_up_bike_dat"))
make_bike_map(dat)
}
