% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{make_plot_dat}
\alias{make_plot_dat}
\title{Make Plot Data from Station and Status}
\usage{
make_plot_dat(station, status)
}
\arguments{
\item{station}{station df}

\item{status}{status df}
}
\value{
combined plot data
}
\description{
Make Plot Data from Station and Status
}
\examples{
make_plot_dat(pins::pin_get("break_up_station_dat"), pins::pin_get("break_up_bike_dat"))
}
